+++ DesignToolbox/DesignToolbox/Pages/Components/Checkbox/CheckboxItem/CheckboxItemConfiguration.swift
@@ -133,48 +133,31 @@ struct CheckboxItemConfiguration: View {
     @Environment(\.theme) private var theme
 
     var body: some View {
-        VStack(alignment: .leading, spacing: theme.spaces.spaceFixedMedium) {
+        VStack(alignment: .leading, spacing: theme.spaces.spaceFixedNone) {
+            OUDSSwitchItem("app_components_checkbox_selection_label", isOn: $model.indicatorState)
 
-            Toggle("app_components_checkbox_selection_label", isOn: $model.indicatorState)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
+            OUDSSwitchItem("app_components_controlItem_icon_label", isOn: $model.icon)
 
-            Toggle("app_components_controlItem_icon_label", isOn: $model.icon)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
-
-            Toggle("app_components_controlItem_flipIcon_label", isOn: $model.flipIcon)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
+            OUDSSwitchItem("app_components_controlItem_flipIcon_label", isOn: $model.flipIcon)
                 .disabled(!model.icon)
 
-            Toggle("app_components_controlItem_divider_label", isOn: $model.divider)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
+            OUDSSwitchItem("app_components_controlItem_divider_label", isOn: $model.divider)
 
-            Toggle("app_components_controlItem_reversed_label", isOn: $model.isReversed)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
+            OUDSSwitchItem("app_components_controlItem_reversed_label", isOn: $model.isReversed)
 
-            Toggle("app_common_enabled_label", isOn: $model.enabled)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
+            OUDSSwitchItem("app_common_enabled_label", isOn: $model.enabled)
                 .disabled(model.isError || model.isReadOnly)
 
-            Toggle("app_components_controlItem_readOnly_label", isOn: $model.isReadOnly)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
+            OUDSSwitchItem("app_components_controlItem_readOnly_label", isOn: $model.isReadOnly)
                 .disabled(!model.enabled || model.isError)
 
-            Toggle("app_components_common_error_label", isOn: $model.isError)
-                .typeHeadingMedium(theme)
-                .oudsForegroundStyle(theme.colors.colorContentDefault)
+            OUDSSwitchItem("app_components_common_error_label", isOn: $model.isError)
                 .disabled(!model.enabled || model.isReadOnly)
+        }
 
-            DesignToolboxEditContentDisclosure {
-                DesignToolboxTextField(text: $model.labelText)
-                DesignToolboxTextField(text: $model.helperText)
-            }
+        DesignToolboxEditContentDisclosure {
+            DesignToolboxTextField(text: $model.labelText)
+            DesignToolboxTextField(text: $model.helperText)
         }
     }
 }